+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 
 ### Fixed
 
-- Fixed memory corruption in SpatialOS Components using more then 16 fields. [#1378](https://github.com/spatialos/gdk-for-unity/pull/1378)
+- Fixed memory corruption in SpatialOS components with more than 16 fields. [#1378](https://github.com/spatialos/gdk-for-unity/pull/1378)
 
 ## `0.3.6` - 2020-05-26